added a few getters and renamed file
