+++ lib/views/pc/team_info/widgets/quick_data/quick_data_card.dart
